MLP: now all is working and having a good performance

+++ GILS-RVND/MLP/benchmark/bm.py
@@ -143,7 +143,7 @@ for line in fin.readlines():
         num = my_list[1]
 
         if flag == 0:
-            sum_cost += int(num)
+            sum_cost += float(num)
             flag = 1
         elif flag == 1:
             sum_time += float(num)
